face_log_server/requirements.txt
# Hikvision Face Log Server - Production Dependencies
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
PyMySQL>=1.1.0
waitress>=2.1.2
python-dotenv>=1.0.0
